pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.3.12",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
